+++ cargo-near-build/Cargo.toml
@@ -9,3 +9,6 @@ camino = "1.1.1"
 colored = "2.0"
 dunce = "1"
 log = "0.4"
+bs58 = "0.5"
+hex = "0.4.3"
+sha2 = "0.10"
