+++ cargo-near/src/lib.rs
@@ -11,9 +11,9 @@ pub mod util;
 pub use build_extended::{
     build as build_extended, BuildScriptOpts, OptsExtended as BuildOptsExtended,
 };
+pub use cargo_near_build::types::near::CompilationArtifact as BuildArtifact;
 pub use commands::build_command::build::run as build;
 pub use commands::build_command::build::Opts as BuildOpts;
-pub use util::CompilationArtifact as BuildArtifact;
 
 #[derive(Debug, Clone, interactive_clap::InteractiveClap)]
 #[interactive_clap(context = near_cli_rs::GlobalContext)]
