+++ cargo-near/src/util/mod.rs
@@ -5,13 +5,13 @@ use std::io::{BufRead, BufReader};
 use std::process::Command;
 use std::thread;
 
-use camino::{Utf8Path, Utf8PathBuf};
+use camino::Utf8Path;
 use cargo_metadata::{Artifact, Message};
 use cargo_near_build::types::cargo::manifest_path::ManifestPath;
+use cargo_near_build::types::near::{CompilationArtifact, VersionMismatch};
 use color_eyre::eyre::{ContextCompat, WrapErr};
 
 use cargo_near_build::types::color_preference::ColorPreference;
-use sha2::{Digest, Sha256};
 
 /// Invokes `cargo` with the subcommand `command`, the supplied `args` and set `env` variables.
 ///
@@ -109,67 +109,6 @@ where
         Ok(result?)
     } else {
         color_eyre::eyre::bail!("`{:?}` failed with exit code: {:?}", cmd, output.code());
-    }
-}
-
-#[derive(Debug)]
-pub enum VersionMismatch {
-    Some {
-        environment: String,
-        current_process: String,
-    },
-    None,
-    UnknownFromDocker,
-}
-
-impl std::fmt::Display for VersionMismatch {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Some {
-                environment,
-                current_process,
-            } => {
-                write!(
-                    f,
-                    "`cargo-near` version {} -> `cargo-near` environment version {}",
-                    current_process, environment
-                )
-            }
-            Self::None => write!(f, "no `cargo-near` version mismatch in nested builds detected",),
-            Self::UnknownFromDocker => write!(f, "it's unknown if `cargo-near` version mismatch has occurred in docker build environment",),
-        }
-    }
-}
-
-pub struct CompilationArtifact {
-    pub path: Utf8PathBuf,
-    pub fresh: bool,
-    pub from_docker: bool,
-    pub cargo_near_version_mismatch: VersionMismatch,
-}
-pub struct SHA256Checksum {
-    hash: Vec<u8>,
-}
-
-impl SHA256Checksum {
-    pub fn to_hex_string(&self) -> String {
-        hex::encode(&self.hash)
-    }
-
-    pub fn to_base58_string(&self) -> String {
-        bs58::encode(&self.hash).into_string()
-    }
-}
-
-impl CompilationArtifact {
-    pub fn compute_hash(&self) -> color_eyre::eyre::Result<SHA256Checksum> {
-        let mut hasher = Sha256::new();
-        hasher.update(std::fs::read(&self.path)?);
-        let hash = hasher.finalize();
-        let hash: &[u8] = hash.as_ref();
-        Ok(SHA256Checksum {
-            hash: hash.to_vec(),
-        })
     }
 }
